+++ core/src/Microsoft.Teams.Apps/Handlers/HtmlWidget/CallToolRequest.cs
