+++ core/src/Microsoft.Teams.Apps/Handlers/HtmlWidget/CallToolResult.cs
