+++ core/src/Microsoft.Teams.Apps/Handlers/HtmlWidget/HtmlWidgetHelpers.cs
