+++ core/src/Microsoft.Teams.Apps/Handlers/HtmlWidget/HtmlWidgetPayload.cs
