+++ core/src/Microsoft.Teams.Apps/Handlers/HtmlWidget/UpdateModelContext.cs
